Move tests from error_handling.rs to corresponding subcommands
