+++ crates/sncast/tests/data/scripts/call/Scarb.toml
@@ -1,8 +1,7 @@
 [package]
-name = "deploy_test_scripts"
+name = "call_test_scripts"
 version = "0.1.0"
 
 [dependencies]
-starknet = ">=2.4.0"
+starknet = ">=2.3.0"
 sncast_std = { path = "../../../../../../sncast_std" }
-
