+++ crates/sncast/tests/data/scripts/call/src/invalid_calldata.cairo
