+++ crates/sncast/tests/data/scripts/call/src/invalid_entry_point.cairo
