+++ crates/sncast/tests/data/scripts/call/src/lib.cairo
