--- crates/sncast/tests/data/scripts/declare/no_contract/Scarb.toml
@@ -1,8 +1,0 @@
-[package]
-name = "no_contract_script"
-version = "0.1.0"
-
-[dependencies]
-starknet = ">=2.3.0"
-sncast_std = { path = "../../../../../../../sncast_std" }
-
