--- crates/sncast/tests/data/scripts/declare/no_contract/src/lib.cairo
@@ -1,2 +1,0 @@
-mod no_contract;
-
