+++ crates/sncast/tests/data/scripts/declare/test_scripts/src/insufficient_account_balance.cairo
