+++ crates/sncast/tests/data/scripts/declare/test_scripts/src/lib.cairo
@@ -1,0 +1,6 @@
+mod no_contract;
+mod same_contract_twice;
+mod with_invalid_max_fee;
+mod with_invalid_nonce;
+mod insufficient_account_balance;
+
