+++ crates/sncast/tests/data/scripts/declare/test_scripts/src/no_contract.cairo
