+++ crates/sncast/tests/data/scripts/declare/test_scripts/src/same_contract_twice.cairo
