+++ crates/sncast/tests/data/scripts/declare/test_scripts/src/with_invalid_max_fee.cairo
