+++ crates/sncast/tests/data/scripts/declare/test_scripts/src/with_invalid_nonce.cairo
