+++ crates/sncast/tests/data/scripts/deploy/src/invalid_calldata.cairo
