+++ crates/sncast/tests/data/scripts/deploy/src/invalid_class_hash.cairo
