+++ crates/sncast/tests/data/scripts/deploy/src/lib.cairo
@@ -1,2 +1,5 @@
 mod with_calldata;
-
+mod same_class_hash_and_salt;
+mod invalid_class_hash;
+mod invalid_calldata;
+mod invalid_nonce;
