+++ crates/sncast/tests/data/scripts/deploy/src/same_class_hash_and_salt.cairo
