--- crates/sncast/tests/data/scripts/error_handling/call/Scarb.toml
@@ -1,7 +1,0 @@
-[package]
-name = "error_handling"
-version = "0.1.0"
-
-[dependencies]
-starknet = ">=2.3.0"
-sncast_std = { path = "../../../../../../../sncast_std" }
