--- crates/sncast/tests/data/scripts/error_handling/declare/src/declare_missing_contract.cairo
@@ -1,13 +1,0 @@
-use sncast_std::{declare, DeclareResult, ScriptCommandError, RPCError, StarknetError};
-
-
-fn main() {
-    let declare_result = declare('Mapaaaa', Option::None, Option::None).unwrap_err();
-    println!("{:?}", declare_result);
-
-    assert(
-        ScriptCommandError::ContractArtifactsNotFound == declare_result,
-        'ohno'
-    )
-}
-
