--- crates/sncast/tests/data/scripts/error_handling/declare/src/lib.cairo
@@ -1,5 +1,0 @@
-mod declare_missing_contract;
-mod declare_same_contract_twice;
-mod declare_with_invalid_max_fee;
-mod declare_with_invalid_nonce;
-mod declare_insufficient_account_balance;
