--- crates/sncast/tests/data/scripts/error_handling/deploy/src/lib.cairo
@@ -1,4 +1,0 @@
-mod deploy_same_class_hash_and_salt;
-mod deploy_invalid_class_hash;
-mod deploy_invalid_calldata;
-mod deploy_invalid_nonce;
