+++ crates/sncast/tests/e2e/script/mod.rs
@@ -1,6 +1,5 @@
 mod call;
 mod declare;
 mod deploy;
-mod error_handling;
 mod general;
 mod invoke;
